Turned off libreadline support on all platforms until the horrible problems experienced on Windows Vista and the lesser but still irritating problems on other platforms can be fixed. Fixed some bugs in the way libreadline was being used, but with libreadline disabled this is a moot point.
Fixed the main GUI (VirtualAGC) so that the choice of folders
in some file-dialogs is better, with hopefully less bleed-
through from one dialog to the other when directories are 
using the dialogs.

Custom AGC/AEA source code can now be symbolically debugged
or be browsed from the main GUI.  Previously, it could only
be assembled or run (in normal mode, without debugging) from
the GUI.  This should make (command line) debugging much 
simpler to do.

Added some mods for compiling yaAGC (only) for the iPhone.

Added some proofing for Luminary 99.