Altered the duty cycle of flashing digits and indicators in the yaDSKY2 program on the basis of feedback from Christian Bucher, from 1:1 (high:low) to 3:1.
Added the contributed AGC game pack from John Pultorak.  In this initial
save, the files are exactly as John sent them.  However, it presently
assembles only with John's assembler, and isn't compatible with yaYUL.
I'll be making whatever changes are necessary to get it to be compatible
with yaYUL, but I wanted to get the originals into the system before
starting.