First attempt at a Docker configuration for a VirtualAgc build.
To use it, you'll need Docker (on Linux, Mac or Windows).
Then do:
 $ docker build -t virtualagc .

This config so far just builds the tools and yaAGC, and compiles
all the mission files and runs a ropediff check on each generatd
rope binary against the known good image, where available.

This is based on an Ubuntu 16.04 base image.

Known issue: the graphical build targets are much harder, with
requirments for various versions of X libs, SDL, Allegro, etc.