yaAGC: Overflow-correct the L register at the end of agc_engine()
The AGC, when reading the L register, overflow-corrected the read value;
thus, the L register cannot be used to store an overflow condition. Per
suggestion from Ron, it was much easier to simply overflow-correct the
register itself at the end of agc_engine(), rather than bothering with
trying to do it on-write or on-read. It's all invisible to the software
running on the AGC anyways.

I've also made changes to Validation to get it to pass again. The
overflow correction is explicitly tested for in ValidateLXCH. Elsewhere,
uses of L to stash an overflow condition have been changed to use Q
instead.