WireLister: Added notion of fan-in/out gates, and fixed a lot of errors
Fan-in outputs are prefixed with an underscore, and fan-out outputs are
prefixed with a caret. It's no longer considered an error for fan-in or
fan-out gates to drive the same signal as that signal's main source. For
multi-module lines like the control pulses, the source was declared to
be the one gate with its VCC pin hooked up. With this and correction of
some missing signals, all "Signals with multiple sources" errors have
been fixed. (Yay!)

I've also greatly reduced the number of signals with no sinks. This
involved filling out the presumed memory module I/O more. I also added a
section for the Test Set/Monitor because, while they are external to the
computer, they are the sink for nearly a hundred signals. There's now
only a rather small amount of this error left, most of which are coming
from the Scaler putting everything onto the backplane despite most of it
not being used.

Lastly, the "signals with no source" was reduced to 1 by assuming that
TO61 == T06/. This, while not explicitly listed as a thing in AGC4 Memo
9, is consistent with the bank registers being stashed in the adder
during Monitor store/fetch sequences.

The only remaining signal without a source is input channel 33, bit 5's
external driver. Suspiciously, there's exactly one AGC input in modules
27-29 that has no sink... I suspect these two are one and the same, and
we missed a rename at some point.