Burned down most of the rest of the "no sink" errors.
There were several real errors mixed in (typos or missed signals), and
those have been fixed. I also decided that nets directly tied to the
outputs of flip-flops should count as 'inout', since it really would
function in that manner if anything were attached to the pin -- due to
the cross-coupled nature they're connected to both an input and an
output.

A few more signals were won by adding in the AGC POWER SUPPLY modules
A30 and A31. They have a lot of other stuff connecting to them than what
I listed (just the logic module connections), but most if it is
power-related, which hasn't been tracked thus far.

All the remaining errors are "real".
* CO02 is broken out to its pins for modularity reasons; nothing uses it.
* FS02A-5A have no sink I can find. I don't know why these gates exist
* MSBSTP is a monitor input with no apparent sink. I think it may have
  hit the analog alarms module or something
* I think OPCDEL and OPMSW2 are probably tied together.
* I've got nothing on Q2A or OUTCOM -- they look like they go to another
  internal module but I've no idea which it would be.