yaAGC+yaDSKY2: Updates to make DSKY handling more 'realistic'
This adds the fictitious output channel 163, whose bits correspond
essentially 1-1 with discretes to the real DSKY. Verb/Noun, KEY REL and
OPER ERR flashing is now timed by the AGC itself (with VN being active
low, and the others being active high).

The RESTART light has also been plumbed through this channel. It is set
by turning the computer on, or getting a GOJAM from a hardware alarm. It
can also be temporarily turned on via the Alarm Test bit in channel 13.
It's extinguished via a reset from channel 10.