Reversed state of metadata correction.
It turns out that the aperture-card scanner is still messing up the card metadata from time to time, in the same way, but I hadn't noticed it since it was less frequent.  So I've made correction the default, with the command-line switch turning it off.  Also, what I had been doing before to correct sheet numbers in which the 2nd digit was 8 (but incorrectly turned into a space by the scanner) was incorrect, so I was getting sheet numbers like "000" and "001" when I should have had "008" and "018".  I'm uncertain whether my fix is correct or not, but I'm taking a shot!